Update to 1 servlet
