+++ src/main/java/org/s2f/mb/service/db/DatabaseHandler.java
@@ -1,4 +1,4 @@
-package org.s2f.mb.model.logic;
+package org.s2f.mb.service.db;
 
 import org.s2f.mb.model.dto.Product;
