+++ src/main/webapp/add-to-fridge.html
@@ -7,7 +7,7 @@
 <body>
 <h1>Add a product to the fridge:</h1>
 <br>
-<form action="AddToFridgeServlet" method="post">
+<form action="AddAndShowServlet" method="post">
     Product name <input type="text" name="name">
     Weight, g <input type="text" name="weight">
     Price, Rub <input type="text" name="price">
@@ -17,6 +17,6 @@
 </form>
 
 <br>
-<a href="ShowAllServlet">Show all products</a>
+<a href="AddAndShowServlet">Show all products</a>
 </body>
 </html>
