--- src/main/webapp/show-all.jsp
@@ -1,14 +1,0 @@
-<%@ page contentType="text/html; charset=UTF-8" language="java" %>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Add to fridge</title>
-</head>
-<body>
-<h1>A list of products</h1>
-<br>
-
-<br>
-<a href="add-to-fridge.html">Add a new product</a>
-</body>
-</html>
